separated filters from tasks view close #23
